Update Ch 5 Ex 2-3 termF() types

+++ ppp_book_practice/Ch 5/ex 2-3.cpp
@@ -99,12 +99,12 @@ double primary() {
 }
 
 double termF() {
-    int left{ (int)primary() };
+    double left{ primary() };
     Token t{ ts.get() };
     while (true) {
         switch (t.kind) {
             case '!':
-                left = factorial(left);
+                left = factorial((int)left);
                 t = ts.get();
                 break;
             default:
